+++ src/Bundle/Core/Tests/Unit/Admin/Extension/PublishTimePeriodExtensionTest.php
@@ -3,12 +3,11 @@
 /*
  * This file is part of the Symfony CMF package.
  *
- * (c) 2011-2013 Symfony CMF
+ * (c) 2011-2014 Symfony CMF
  *
  * For the full copyright and license information, please view the LICENSE
  * file that was distributed with this source code.
  */
-
 
 namespace Symfony\Cmf\Bundle\CoreBundle\Tests\Unit\Admin\Extension;
 
@@ -22,13 +21,14 @@ class PublishTimePeriodExtensionTest extends \PHPUnit_Framework_Testcase
             'Sonata\AdminBundle\Form\FormMapper'
         )->disableOriginalConstructor()->getMock();
 
-        $this->extension = new PublishTimePeriodExtension();
+        $this->extension = new PublishTimePeriodExtension('some_group');
     }
 
     public function testFormMapper()
     {
         $this->formMapper->expects($this->once())
             ->method('with')
+            ->with('some_group')
             ->will($this->returnSelf());
         $this->formMapper->expects($this->exactly(2))
             ->method('add')