+++ src/Bundle/Core/Tests/Unit/Twig/Extension/CmfExtensionTest.php
@@ -3,12 +3,11 @@
 /*
  * This file is part of the Symfony CMF package.
  *
- * (c) 2011-2013 Symfony CMF
+ * (c) 2011-2014 Symfony CMF
  *
  * For the full copyright and license information, please view the LICENSE
  * file that was distributed with this source code.
  */
-
 
 namespace Symfony\Cmf\Bundle\CoreBundle\Tests\Unit\Twig\Extension;
 
@@ -16,6 +15,17 @@ use Symfony\Cmf\Bundle\CoreBundle\Twig\Extension\CmfExtension;
 
 class CmfExtensionTest extends \PHPUnit_Framework_TestCase
 {
+    private $cmfHelper;
+    /**
+     * @var \Twig_Environment
+     */
+    private $env;
+
+    /**
+     * @var CmfExtension
+     */
+    private $cmfExtension;
+
     public function setUp()
     {
         $this->cmfHelper = $this->getMockBuilder(
@@ -30,6 +40,6 @@ class CmfExtensionTest extends \PHPUnit_Framework_TestCase
     public function testFunctions()
     {
         $functions = $this->cmfExtension->getFunctions();
-        $this->assertCount(15, $functions);
+        $this->assertCount(16, $functions);
     }
 }