+++ src/Bundle/Core/Translatable/TranslatableInterface.php
@@ -3,12 +3,11 @@
 /*
  * This file is part of the Symfony CMF package.
  *
- * (c) 2011-2013 Symfony CMF
+ * (c) 2011-2014 Symfony CMF
  *
  * For the full copyright and license information, please view the LICENSE
  * file that was distributed with this source code.
  */
-
 
 namespace Symfony\Cmf\Bundle\CoreBundle\Translatable;
 
@@ -22,13 +21,13 @@ interface TranslatableInterface
 {
     /**
      * @return string|boolean The locale of this model or false if
-     *      translations are disabled in this project.
+     *                        translations are disabled in this project.
      */
     public function getLocale();
 
     /**
      * @param string|boolean $locale The local for this model, or false if
-     *      translations are disabled in this project.
+     *                               translations are disabled in this project.
      */
     public function setLocale($locale);
 }